refactor: move NotificationPreferences and UpdateNotificationPreferencesParams to notification.types module
