+++ src/apps/backend/modules/notification/types.py
@@ -16,6 +16,13 @@ class EmailRecipient:
 
 
 @dataclass(frozen=True)
+class NotificationPreferences:
+    email_enabled: bool = True
+    push_enabled: bool = True
+    sms_enabled: bool = True
+
+
+@dataclass(frozen=True)
 class SendEmailParams:
     recipient: EmailRecipient
     sender: EmailSender
@@ -33,6 +40,14 @@ class SendSMSParams:
 class CommunicationErrorCode:
     VALIDATION_ERROR = "COMMUNICATION_ERR_01"
     SERVICE_ERROR = "COMMUNICATION_ERR_02"
+
+
+@dataclass(frozen=True)
+class UpdateNotificationPreferencesParams:
+    account_id: str
+    email_enabled: bool
+    push_enabled: bool
+    sms_enabled: bool
 
 
 @dataclass(frozen=True)
